Move DirectoryDB from TestingState to Minithesis

+++ src/main/java/com/justinblank/minithesis/Minithesis.java
@@ -2,6 +2,7 @@ package com.justinblank.minithesis;
 
 import org.opentest4j.AssertionFailedError;
 
+import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Random;
 import java.util.function.Consumer;
@@ -11,18 +12,27 @@ public class Minithesis {
 
     public static <T> void runTest(Consumer<TestCase> test, String name) {
         Function<TestCase, TestResult<T>> testFunction = wrapConsumer(test);
+        DirectoryDB db = null;
+        try {
+            db = new DirectoryDB(Path.of("target/minithesis-cache"));
+        }
+        catch (Exception e) {
+            e.printStackTrace();
+        }
         var state = new TestingState<>(new Random(), testFunction, 100);
 
-        runTest(test, name, testFunction, state);
+        runTest(test, name, testFunction, state, db);
     }
 
-    static <T> void runTest(Consumer<TestCase> test, String name, Function<TestCase, TestResult<T>> testFn, TestingState<T> state) {
-        var db = state.getDB();
-        var previous_failure = db.get(name);
-        if (previous_failure != null) {
-            var tc = TestCase.forChoices(previous_failure, false);
-            state.applyTestFunction(tc);
+    static <T> void runTest(Consumer<TestCase> test, String name, Function<TestCase, TestResult<T>> testFn, TestingState<T> state, DirectoryDB db) {
+        if (db != null) {
+            var previous_failure = db.get(name);
+            if (previous_failure != null) {
+                var tc = TestCase.forChoices(previous_failure, false);
+                state.applyTestFunction(tc);
+            }
         }
+
         if (state.getResult() == null) {
             state.run();
         }
@@ -31,14 +41,18 @@ public class Minithesis {
             throw new UnsatisfiableTestCaseException();
         }
         if (state.getResult() == null) {
-            db.delete(name);
+            if (db != null) {
+                db.delete(name);
+            }
         } else {
             int[] choices = new int[state.getResult().size()];
             for (int i = 0; i < state.getResult().size(); i++) {
                 choices[i] = state.getResult().get(i);
             }
 
-            db.set(name, choices);
+            if (db != null) {
+                db.set(name, choices);
+            }
 
             var testCase = TestCase.forChoices(new ArrayList<>(state.getResult()), false);
             test.accept(testCase);
